# [core] Speed up and deflake `test_output.py`